update measure en cours + gestion probleme date en cours
