+++ src/app/(main)/(dashboard)/weightGoalForm.tsx
@@ -1,10 +1,11 @@
 import { createWeightGoalAction, updateWeightGoalAction } from "@/app/actions/weightGoalActions";
+import { WeightGoalActionState } from "@/libs/types/actionState";
 import { Input } from "@/ui/shadcn/components/ui/input";
 import { Label } from "@/ui/shadcn/components/ui/label";
 import clsx from "clsx";
 import { useActionState } from "react";
 
-const initialState = {
+const initialState: WeightGoalActionState = {
     status: "",
     message: "",
 };
