+++ src/app/(main)/mesures/page.tsx
@@ -1,5 +1,5 @@
 import { getServerAuthSession } from "@/libs/server/nextAuthSession";
-import AddMeasuresForm from "./addMeasuresForm";
+import AddMeasurementForm from "./addMeasurementForm";
 import { redirect } from "next/navigation";
 import { getMeasurementsByUser } from "@/libs/server/database/measurement";
 import { Measurement } from "@/libs/types/measurement";
@@ -22,7 +22,7 @@ export default async function MesuresPage() {
     return (
         <div className="pt-10">
             <h1 className="mb-10 text-4xl font-poppins font-medium">Mes mesures</h1>
-            <AddMeasuresForm latestMeasurement={measurements[0]} />
+            <AddMeasurementForm latestMeasurement={measurements[0]} />
             <MeasurementsHistory measurements={measurements} />
         </div>
     );
