+++ src/libs/server/database/measurement.ts
@@ -17,12 +17,24 @@ export async function getMeasurementsByUser(userId: string): Promise<Measurement
     return (rows as MeasurementDb[]).map(mapMeasurementDbToMeasurement);
 }
 
-export async function getUserMeasurementByDate(userId: string, date: Date): Promise<null | Measurement> {
+export async function getMeasurementByDate(userId: string, date: Date): Promise<null | Measurement> {
     const measuredAtString = date.toISOString().slice(0, 19).replace("T", " ");
     const [rows] = await pool.query<RowDataPacket[]>(
         "SELECT * FROM measurement WHERE user_id = ? AND measured_at = ?",
         [userId, measuredAtString]
     );
+    const measurement = rows[0] as MeasurementDb;
+    if (!measurement) {
+        return null;
+    }
+    return mapMeasurementDbToMeasurement(measurement);
+}
+
+export async function getMeasurementById(userId: string, measurementId: string): Promise<null | Measurement> {
+    const [rows] = await pool.query<RowDataPacket[]>("SELECT * FROM measurement WHERE user_id = ? AND id = ?", [
+        userId,
+        measurementId,
+    ]);
     const measurement = rows[0] as MeasurementDb;
     if (!measurement) {
         return null;
@@ -44,7 +56,7 @@ export async function createMeasurement(
         arm: number | null;
     }
 ) {
-    const measuredAtString = data.measuredAt.toISOString().slice(0, 19).replace("T", " ");
+    const measuredAtString = data.measuredAt.toISOString().slice(0, 10);
     await pool.query<ResultSetHeader>(
         "INSERT INTO measurement (user_id, measured_at, weight, chest, underbust, waist, belly, hips, thigh, arm) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?);",
         [
@@ -58,6 +70,35 @@ export async function createMeasurement(
             data.hips,
             data.thigh,
             data.arm,
+        ]
+    );
+}
+
+export async function updateMeasurement(
+    measurementId: string,
+    data: {
+        weight: number | null;
+        chest: number | null;
+        underbust: number | null;
+        waist: number | null;
+        belly: number | null;
+        hips: number | null;
+        thigh: number | null;
+        arm: number | null;
+    }
+) {
+    await pool.query<ResultSetHeader>(
+        "UPDATE measurement SET weight = ?, chest = ?, underbust = ?, waist = ?, belly = ?, hips = ?, thigh = ?, arm = ? WHERE id = ?",
+        [
+            data.weight,
+            data.chest,
+            data.underbust,
+            data.waist,
+            data.belly,
+            data.hips,
+            data.thigh,
+            data.arm,
+            measurementId,
         ]
     );
 }
